restructred utilities
