--- Code/ABSdata.csv
@@ -1,21 +1,0 @@
-X,T
--0.250919762305275,0.250919762305275
-0.9014286128198323,0.9014286128198323
-0.4639878836228102,0.4639878836228102
-0.1973169683940732,0.1973169683940732
--0.687962719115127,0.687962719115127
--0.6880109593275947,0.6880109593275947
--0.8838327756636011,0.8838327756636011
-0.7323522915498704,0.7323522915498704
-0.2022300234864176,0.2022300234864176
-0.416145155592091,0.416145155592091
--0.9588310114083951,0.9588310114083951
-0.9398197043239886,0.9398197043239886
-0.6648852816008435,0.6648852816008435
--0.5753217786434477,0.5753217786434477
--0.6363500655857988,0.6363500655857988
--0.6331909802931324,0.6331909802931324
--0.39151551408092455,0.39151551408092455
-0.04951286326447568,0.04951286326447568
--0.13610996271576847,0.13610996271576847
--0.4175417196039162,0.4175417196039162
