+++ Code/LateResidualConnections.py
@@ -1,14 +1,17 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# Personal Files Import 
-import LateResidualNeuralNetwork
-import LateResidualUtilityFunctions as lr_utils
-
 import time
 import pandas as pd
 import numpy as np
 from tqdm import tqdm
+
+# Personal Files Import 
+import LateResidualPyTorch.LateResidualNeuralNetwork as LateResidualNeuralNetwork
+import Utilities.LateResidualUtilityFunctions as lr_utils
+import Utilities.GraphingCode.LateResidualGraphing as graph_utils
+import Utilities.DataframeCode.LateResidualDataframe as dataframe_utils
+
 
 def run_experiment(X, T, epochs, network_architecture, optimizer, learning_rate, connection_style, training_style, verbose=False):
 
@@ -32,7 +35,7 @@ def run_experiment(X, T, epochs, network_architecture, optimizer, learning_rate,
 
 
 def run_experiments(optimizers, learning_rates, network_architectures, connection_styles, training_styles, iterations, epochs, width, depths):
-    X, T = lr_utils.load_abs_data()
+    X, T = dataframe_utils.load_abs_data()
 
     dataframe_column_names = ['Iterative - Total Converged', 'Iterative - Amount of Dead Neurons', 'Iterative - Amount of Dead Layers', 'Iterative - Total Time', 
                                     'Batch - Total Converged', 'Batch - Amount of Dead Neurons', 'Batch - Amount of Dead Layers', 'Batch - Total Time']
@@ -60,7 +63,7 @@ def run_experiments(optimizers, learning_rates, network_architectures, connectio
                         for training_style in training_styles: 
                             for iteration in tqdm(range(iterations)):
                                 did_converge, results, model = run_experiment(X, T, epochs, network_architecture, optimizer, learning_rate, connection_style, training_style)
-                                lr_utils.graph_results(model, learning_rate, network_architecture, width, optimizer, iteration, training_style, did_converge)
+                                graph_utils.graph_results(model, learning_rate, network_architecture, width, optimizer, iteration, training_style, did_converge)
                                 
                                 if did_converge:
                                     if training_style == 'Iterative':
@@ -75,9 +78,9 @@ def run_experiments(optimizers, learning_rates, network_architectures, connectio
                         
                 dataframe.replace(to_replace=np.nan, value=' ', inplace=True)
                 dataframe.insert(len(dataframe.columns), ' ', [' ' for _ in range(len(depths) * 2 + 1)], True)
-                lr_utils.save_dataframe_to_csv(dataframe, width, optimizer, learning_rate)
+                dataframe_utils.save_dataframe_to_csv(dataframe, width, optimizer, learning_rate)
                 
-    lr_utils.combine_all_dataframes_to_csv(width, optimizers)
+    dataframe_utils.combine_all_dataframes_to_csv(width, optimizers)
     lr_utils.print_end_of_all_training_message()
